rubygems: inject proper compiler flags
by patching the cross rbconfig.rb on the fly.
also introduce properly version so-files.
as the ruby dynlinker only seems to handle plain .so
files correctly create symlinks and package them to the
base package.
remove the moving of shared objects to the site-dir, as
this isn't needed anymore

This patch removes the need to INSANE_SKIP ldflags and dev-elf

Relates to #20

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>