CI: use github.sha as unique marker for cron/push
and github.base_ref for the pull request pipelines.
As the base_ref information can be a name or a sha, we will
forcefully translate it to a sha and inject it into the runner
environment (and that forceful part involves a lot of brute force).
This means also that we can't remove git information from the layer.
Remove restore-keys feature to avoid old caches being used.

Closes to #57

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>