ci: move restore key to the right pipeline
as we don't need to restore caches in the push pipeline, as this is the
one that creates those, but in the PR pipelines.

Also remove the custom script to determine the correct
base-sha of the branch and the need to have the git info
in the build workspace.
Switch back to github.base_ref + restore-keys

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>