pkgupdateskip: prevent jekyll upgrade
latest **point** release of jekyll injects a architecture
specific gem 'sass-converter' which is available only for
arm, arm64 and x64 - additionally this gem tries to pull
stuff from the internet while compiling.
This is simply not the way bitbake build system operates.

Prevent any further updates to the gems jekyll and
jekyll-sass-converter.

Closes #457
Closes #438

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>